+++ packages/draft-editor/src/buttons/anchor.tsx
@@ -1,46 +1,51 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { EditorState, RichUtils } from 'draft-js'
-import { ANCHOR_ENTITY_TYPE } from '@kids-reporter/draft-renderer'
+import { ENTITY } from '@kids-reporter/draft-renderer'
+import { AnchorIDEditor } from '../entity-decorators/anchor'
 
 type AnchorButtonProps = {
   className?: string
   isActive: boolean
   editorState: EditorState
   onChange: (arg0: EditorState) => void
+  onEditStart: () => void
+  onEditFinish: () => void
 }
 
 export const AnchorButton = (props: AnchorButtonProps) => {
   const toggleEntity = RichUtils.toggleLink
   const { isActive, editorState, onChange } = props
+  const [isEditorOpen, setIsEditorOpen] = useState(false)
 
   const promptForAnchor = (e: React.MouseEvent<HTMLDivElement>) => {
     e.preventDefault()
-
     const selection = editorState.getSelection()
     if (!selection.isCollapsed()) {
-      const contentState = editorState.getCurrentContent()
-      const startKey = selection.getStartKey()
-      const block = contentState.getBlockForKey(startKey)
-      const start = selection.getStartOffset()
-      const end = selection.getEndOffset()
-      const selectedText = block.getText().slice(start, end)
-      const contentStateWithEntity = contentState.createEntity(
-        ANCHOR_ENTITY_TYPE,
-        'IMMUTABLE',
-        {
-          anchorKey: `createdAt-${Date.now()}`,
-          anchorLabel: selectedText,
-        }
-      )
-      const entityKey = contentStateWithEntity.getLastCreatedEntityKey()
-      const newEditorState = EditorState.set(editorState, {
-        currentContent: contentStateWithEntity,
-      })
-
-      onChange(
-        toggleEntity(newEditorState, newEditorState.getSelection(), entityKey)
-      )
+      props.onEditStart()
+      setIsEditorOpen(true)
     }
+  }
+
+  const confirmAnchor = (anchorID: string) => {
+    const contentState = editorState.getCurrentContent()
+    const contentStateWithEntity = contentState.createEntity(
+      ENTITY.Anchor,
+      'MUTABLE',
+      {
+        anchorID: anchorID,
+      }
+    )
+    const entityKey = contentStateWithEntity.getLastCreatedEntityKey()
+    const newEditorState = EditorState.set(editorState, {
+      currentContent: contentStateWithEntity,
+    })
+
+    onChange(
+      toggleEntity(newEditorState, newEditorState.getSelection(), entityKey)
+    )
+
+    setIsEditorOpen(false)
+    props.onEditFinish()
   }
 
   const removeAnchor = () => {
@@ -48,15 +53,27 @@ export const AnchorButton = (props: AnchorButtonProps) => {
     if (!selection.isCollapsed()) {
       onChange(toggleEntity(editorState, selection, null))
     }
+    setIsEditorOpen(false)
+    props.onEditFinish()
   }
 
   return (
-    <div
-      className={props.className}
-      onMouseDown={isActive ? removeAnchor : promptForAnchor}
-      title="索引"
-    >
-      <span>索</span>
-    </div>
+    <>
+      {isEditorOpen && (
+        <AnchorIDEditor
+          isOpen={isEditorOpen}
+          anchorIDValue={''}
+          onConfirm={confirmAnchor}
+          onCancel={removeAnchor}
+        />
+      )}
+      <div
+        className={props.className}
+        onMouseDown={isActive ? removeAnchor : promptForAnchor}
+        title="錨點"
+      >
+        <span>錨</span>
+      </div>
+    </>
   )
 }