+++ packages/draft-renderer/src/entity-decorators/index.ts
@@ -1,21 +1,14 @@
 import { CompositeDecorator } from 'draft-js'
 import { annotationDecorator } from './annotation-decorator'
 import { linkDecorator } from './link-decorator'
-import {
-  anchorDecorator,
-  findAnchorEntities,
-  ANCHOR_ENTITY_TYPE,
-} from './anchor'
+import { tocAnchorDecorator } from './toc-anchor'
+import { anchorDecorator } from './anchor'
 
-export {
-  annotationDecorator,
-  linkDecorator,
-  findAnchorEntities,
-  ANCHOR_ENTITY_TYPE,
-}
+export { annotationDecorator, linkDecorator }
 
 export const decorator = new CompositeDecorator([
   annotationDecorator,
   linkDecorator,
+  tocAnchorDecorator,
   anchorDecorator,
 ])