+++ packages/draft-renderer/src/entity-decorators/link-decorator.tsx
@@ -1,45 +1,49 @@
 import React from 'react'
-import styled from 'styled-components'
-import { ContentBlock, ContentState } from 'draft-js'
+import styled, { useTheme } from 'styled-components'
+import { ContentState } from 'draft-js'
+import { ENTITY, findEntitiesByType } from '../utils/entity'
 
 const LinkWrapper = styled.a`
   text-decoration: underline;
   color: #27b5f7;
   transition: color 0.1s ease-in;
+  cursor: pointer;
 
   &:hover {
     color: #232323;
   }
 `
 
-function findLinkEntities(
-  contentBlock: ContentBlock,
-  callback: (start: number, end: number) => void,
-  contentState: ContentState
-) {
-  contentBlock.findEntityRanges((character) => {
-    const entityKey = character.getEntity()
-    return (
-      entityKey !== null &&
-      contentState.getEntity(entityKey).getType() === 'LINK'
-    )
-  }, callback)
-}
-
-export const linkDecorator = {
-  strategy: findLinkEntities,
-  component: Link,
-}
-
-function Link(props: {
+const Link = (props: {
   contentState: ContentState
   entityKey: string
   children: React.ReactNode
-}) {
+}) => {
   const { url } = props.contentState.getEntity(props.entityKey).getData()
-  return (
-    <LinkWrapper href={url} target="_blank">
-      {props.children}
-    </LinkWrapper>
-  )
+  const theme = useTheme()
+
+  // Handling for internal/external links, internal links start with '#'
+  const linkProps = url.match(/^#/)
+    ? {
+        onClick: () => {
+          const anchor = document.querySelector(url) as HTMLElement
+          if (anchor) {
+            window.scrollTo({
+              top: anchor.offsetTop - theme?.offsetTop,
+              behavior: 'smooth',
+            })
+          }
+        },
+      }
+    : {
+        href: url,
+        target: '_blank',
+      }
+
+  return <LinkWrapper {...linkProps}>{props.children}</LinkWrapper>
+}
+
+export const linkDecorator = {
+  strategy: findEntitiesByType(ENTITY.Link),
+  component: Link,
 }