+++ packages/frontend/src/app/(sticky-header)/article/[slug]/hero-image.tsx
@@ -7,19 +7,21 @@ type HeroImageProp = {
 }
 
 export const HeroImage = (props: HeroImageProp) => {
-  const image = props?.image // TODO: placeholder for missing image
-  const caption = props?.caption ?? ''
+  const image = props.image // TODO: placeholder for missing image
+  const caption = props.caption ?? ''
   return (
-    <figure className="hero-image">
-      <div className="image-container">
-        <img
-          srcSet={`${image?.resized?.small} 800w, ${image?.resized?.medium} 1200w, ${image?.resized?.large} 2000w"`}
-          sizes="(min-width: 1100px) 1000px, 90vw"
-          src={image?.resized?.medium}
-        />
-      </div>
-      <figcaption>{caption}</figcaption>
-    </figure>
+    image && (
+      <figure className="hero-image">
+        <div className="image-container">
+          <img
+            srcSet={`${image.resized?.small} 320w, ${image.resized?.medium} 500w, ${image.resized?.large} 1000w`}
+            sizes="(min-width: 1100px) 1000px, 90vw"
+            src={image.resized?.medium}
+          />
+        </div>
+        <figcaption>{caption}</figcaption>
+      </figure>
+    )
   )
 }