+++ packages/core/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @kids-reporter/draft-editor@1.0.6
+
+## 1.0.5
+
+### Patch Changes
+
+- @kids-reporter/draft-editor@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- @kids-reporter/draft-editor@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes