+++ packages/draft-editor/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## 1.0.6
+
+### Patch Changes
+
+- Add id validator to anchor
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @kids-reporter/draft-renderer@1.0.5
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @kids-reporter/draft-renderer@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes